refactor(backend): consolidate duplicate venice ai client into single service
