+++ backend/src/services/venice/singleClient.test.ts
@@ -1,0 +1,1 @@
+import fs from 'fs';`nimport path from 'path';`n`ndescribe('VeniceClient Single Canonical Service Guard', () => {`n  it('should ensure obsolete backend/src/venice directory does not exist', () => {`n    const obsoletePath = path.resolve(__dirname, '../../venice');`n    expect(fs.existsSync(obsoletePath)).toBe(false);`n  });`n});
