--- backend/src/venice/client.ts
@@ -1,7 +1,0 @@
-export { VeniceClient } from '../services/venice/client.js';
-export type {
-  AgentType,
-  CompleteOptions,
-  VeniceClientConfig,
-  VeniceClientLike,
-} from '../services/venice/types.js';
