--- backend/src/venice/index.ts
@@ -1,9 +1,0 @@
-export { VeniceClient } from '../services/venice/client.js';
-export type {
-  AgentType,
-  CompleteOptions,
-  VeniceClientConfig,
-  VeniceClientLike,
-} from '../services/venice/types.js';
-export { CircuitBreaker, type CircuitState } from './circuitBreaker.js';
-export { CircuitOpenError, TokenBudgetExceededError } from './errors.js';
